RHOAIENG-79876: Move Prometheus, TrustyAI, and Bias metrics types to model-serving shared
Moves domain-specific types, enums, and constants from frontend into
`packages/model-serving/src/shared/types.ts` as a clean lift-and-delete.
All consumers updated to import from `@odh-dashboard/model-serving/shared/types`.
