+++ frontend/src/__mocks__/mockPrometheusBias.ts
@@ -1,5 +1,6 @@
 /* eslint-disable camelcase */
-import { PrometheusQueryRangeResponse, PrometheusQueryRangeResponseDataResult } from '#~/types';
+import type { PrometheusQueryRangeResponseDataResult } from '@odh-dashboard/model-serving/shared/types';
+import { PrometheusQueryRangeResponse } from '#~/types';
 
 type MockPrometheusServingType = {
   result?: PrometheusQueryRangeResponseDataResult[];
