+++ frontend/src/api/prometheus/kservePerformanceMetrics.ts
@@ -1,13 +1,14 @@
 import React from 'react';
 import { SupportedArea, useIsAreaAvailable } from '@odh-dashboard/plugin-core/areas';
 import { TimeframeTitle } from '@odh-dashboard/ui-core/types/metrics';
+import type { PrometheusQueryRangeResultValue } from '@odh-dashboard/model-serving/shared/types';
 import {
   KserveMetricGraphDefinition,
   NimMetricGraphDefinition,
 } from '#~/concepts/metrics/kserve/types';
 import { defaultResponsePredicate } from '#~/api/prometheus/usePrometheusQueryRange';
 import useQueryRangeResourceData from '#~/api/prometheus/useQueryRangeResourceData';
-import { PendingContextResourceData, PrometheusQueryRangeResultValue } from '#~/types';
+import { PendingContextResourceData } from '#~/types';
 import { DEFAULT_PENDING_CONTEXT_RESOURCE } from '#~/api/prometheus/const';
 
 type RequestCountData = {
