+++ frontend/src/concepts/trustyai/content/useTrustyCRState.tsx
@@ -8,9 +8,9 @@ import {
   SplitItem,
 } from '@patternfly/react-core';
 import type { ProjectKind } from '@odh-dashboard/k8s-core';
+import { TrustyInstallState } from '@odh-dashboard/model-serving/shared/types';
 import useManageTrustyAICR from '#~/concepts/trustyai/useManageTrustyAICR';
 import { useTrustySettingsAccessAllowed } from '#~/concepts/trustyai/content/useTrustySettingsAccessAllowed';
-import { TrustyInstallState } from '#~/concepts/trustyai/types';
 import TrustyAIInstalledState from '#~/concepts/trustyai/content/statusStates/TrustyAIInstalledState';
 import TrustyAIUninstalledState from '#~/concepts/trustyai/content/statusStates/TrustyAIUninstalledState';
