+++ frontend/src/concepts/trustyai/context/useFetchBiasMetricConfigs.ts
@@ -5,8 +5,8 @@ import useFetchState, {
   FetchStateCallbackPromise,
   NotReadyError,
 } from '@odh-dashboard/ui-core/hooks/useFetchState';
+import type { BiasMetricConfig } from '@odh-dashboard/model-serving/shared/types';
 import { TrustyAPIState } from '#~/concepts/trustyai/useTrustyAIAPIState';
-import { BiasMetricConfig } from '#~/concepts/trustyai/types';
 import { formatListResponse } from '#~/concepts/trustyai/utils';
 
 const useFetchBiasMetricConfigs = (apiState: TrustyAPIState): FetchState<BiasMetricConfig[]> => {
