+++ frontend/src/types.ts
@@ -12,6 +12,7 @@ import type {
   PodContainer,
   Volume,
 } from '@odh-dashboard/k8s-core';
+import type { PrometheusQueryRangeResponseDataResult } from '@odh-dashboard/model-serving/shared/types';
 import { FeatureFlag } from '@odh-dashboard/plugin-core/areas';
 import { FetchStateObject } from '@odh-dashboard/ui-core/hooks/useFetch';
 import { HardwarePodSpecOptions } from '#~/concepts/hardwareProfiles/types';
@@ -39,13 +40,6 @@ export type PrometheusQueryResponse<TResultExtraProps extends object = object> =
   status: string;
 };
 
-export type PrometheusQueryRangeResponseDataResult = {
-  metric: {
-    request?: string;
-    pod?: string;
-  };
-  values: PrometheusQueryRangeResultValue[];
-};
 export type PrometheusQueryRangeResponseData = {
   result?: PrometheusQueryRangeResponseDataResult[];
   resultType: string;
@@ -54,8 +48,6 @@ export type PrometheusQueryRangeResponse = {
   data: PrometheusQueryRangeResponseData;
   status: string;
 };
-
-export type PrometheusQueryRangeResultValue = [number, string];
 
 export type NotebookControllerUserState = {
   user: string;
