+++ packages/model-serving/src/components/metrics/bias/BiasConfigurationPage/BiasConfigurationModal/ManageBiasConfigurationModal.tsx
@@ -9,10 +9,9 @@ import {
   ModalFooter,
 } from '@patternfly/react-core';
 import DashboardHelpTooltip from '@odh-dashboard/ui-core/components/DashboardHelpTooltip';
-import { BiasMetricConfig } from '@odh-dashboard/internal/concepts/trustyai/types';
-import { BiasMetricType } from '@odh-dashboard/internal/api';
 import { TrustyAIContext } from '@odh-dashboard/internal/concepts/trustyai/context/TrustyAIContext';
 import DashboardModalFooter from '@odh-dashboard/ui-core/components/DashboardModalFooter';
+import { BiasMetricConfig, BiasMetricType } from '@odh-dashboard/model-serving/shared/types';
 import type { InferenceServiceKind } from '@odh-dashboard/model-serving/shared';
 import useBiasConfigurationObject from './useBiasConfigurationObject';
 import MetricTypeField from './MetricTypeField';
