+++ packages/model-serving/src/components/metrics/bias/BiasConfigurationPage/BiasConfigurationTable.tsx
@@ -4,8 +4,8 @@ import { DashboardEmptyTableView, Table } from '@odh-dashboard/ui-core';
 import DashboardSearchField, {
   SearchType,
 } from '@odh-dashboard/ui-core/components/DashboardSearchField';
-import { BiasMetricConfig } from '@odh-dashboard/internal/concepts/trustyai/types';
 import { useModelBiasData } from '@odh-dashboard/internal/concepts/trustyai/context/useModelBiasData';
+import { BiasMetricConfig } from '@odh-dashboard/model-serving/shared/types';
 import type { InferenceServiceKind } from '@odh-dashboard/model-serving/shared';
 import DeleteBiasConfigurationModal from './BiasConfigurationModal/DeleteBiasConfigurationModal';
 import ManageBiasConfigurationModal from './BiasConfigurationModal/ManageBiasConfigurationModal';
