+++ packages/model-serving/src/components/metrics/bias/useBiasChartSelections.ts
@@ -1,7 +1,7 @@
 import { useParams } from 'react-router-dom';
 import React from 'react';
 import { useBrowserStorage } from '@odh-dashboard/ui-core/utilities';
-import { BiasMetricConfig } from '@odh-dashboard/internal/concepts/trustyai/types';
+import { BiasMetricConfig } from '@odh-dashboard/model-serving/shared/types';
 import { byId } from '../utils';
 
 const SELECTED_CHARTS_STORAGE_KEY_PREFIX = 'odh.dashboard.xai.selected_bias_charts';
