+++ packages/model-serving/src/components/metrics/const.ts
@@ -1,4 +1,4 @@
-import { BiasMetricType } from '@odh-dashboard/internal/api';
+import { BiasMetricType } from '@odh-dashboard/model-serving/shared/types';
 import { BiasChartConfigMap, MetricsChartTypes } from './types';
 import { ModelMetricType } from './ModelServingMetricsContext';
 import { calculateThresholds } from './utils';
