+++ packages/model-serving/src/components/metrics/types.ts
@@ -1,9 +1,11 @@
 import { DomainTuple, ForAxes } from 'victory-core';
 import { ChartThemeDefinitionInterface } from '@patternfly/react-charts/victory';
 import { FetchStateObject } from '@odh-dashboard/ui-core/hooks/useFetch';
-import { PrometheusQueryRangeResultValue } from '@odh-dashboard/internal/types';
-import { BiasMetricType } from '@odh-dashboard/internal/api';
-import { BiasMetricConfig } from '@odh-dashboard/internal/concepts/trustyai/types';
+import {
+  PrometheusQueryRangeResultValue,
+  BiasMetricType,
+  BiasMetricConfig,
+} from '@odh-dashboard/model-serving/shared/types';
 import { ModelMetricType } from './ModelServingMetricsContext';
 
 export type TranslatePoint = (line: GraphMetricPoint) => GraphMetricPoint;
